chore(typing): remove some type checking/coersion and move/delete files
